backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
django-cors-headers==4.3.1
reportlab==4.0.7
openpyxl==3.1.2
djongo==1.3.6
pymongo==3.12.3
dnspython==2.6.1
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
